Continuation of my work after Assignment Submission
made advancements on the model class. now shader and rendering information is stored in the model and is automatically used when rendering
i also made an "Autorenderer" which is a vectorof models. when a model is added to this list, they will be automatically rendered using the autorenderer.

for advanced shading like cellShading, the default renderer (mostly hardcoded) must be used